pyproject.toml
# Robot Dev Team Project
# File: pyproject.toml
# Description: Project metadata, dependencies, and build configuration.
# License: MIT
# SPDX-License-Identifier: MIT
# Copyright (c) 2025 MCKNLY LLC

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "robot-dev-team"
version = "0.1.0"
description = "GitLab webhook listener and LLM agent router"
authors = [{name = "Robot Dev Team"}]
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.135.1",
  "uvicorn[standard]==0.41.0",
  "pydantic-settings==2.13.1",
  "pyyaml==6.0.3",
]

[project.optional-dependencies]
dev = [
  "mypy==1.19.1",
  "pytest==9.0.2",
  "httpx==0.28.1",
  "pytest-asyncio==1.3.0",
  "ruff==0.15.6",
  "types-PyYAML==6.0.12.20250915",
]

[tool.setuptools.packages.find]
where = ["app"]
